refactor(ConfigProvider): split files
